+++ src/main/java/kitchenpos/products/tobe/application/ProductPriceValidator.kt
@@ -1,0 +1,15 @@
+package kitchenpos.products.tobe.application
+
+import org.springframework.stereotype.Component
+import java.math.BigDecimal
+
+@Component
+class ProductPriceValidator {
+    fun requireNormalPrice(
+        price: BigDecimal?,
+    ) = require(price != null && isNotNegativePrice(price))
+
+    private fun isNotNegativePrice(
+        price: BigDecimal,
+    ) = price >= BigDecimal.ZERO
+}
